refactor: separate pingpong files
