+++ backend-ping-pong/src/duel/Rectangle.ts
@@ -1,0 +1,13 @@
+export class Rectangle {
+    x: number;
+    y: number;
+    width: number;
+    height: number;
+
+    constructor(x: number, y: number, width: number, height: number) {
+        this.x = x;
+        this.y = y;
+        this.width = width;
+        this.height = height;
+    }
+}
